Modified when found candidates is set to true
The variable found_candidates_ is now set to true during inside
feedbackCallback(). This gives time for the candidates and their
costs to transfer to the local containers. When found_candidates_ was
set to true is the doneCallback(), compute() would terminate before
the candidate messages were received. This caused the planner to fail.
However, now planning does not fail.